refactored logic of resolving TCP connection info
